+++ src/main/java/de/maxhenkel/camerautils/gui/ThirdPersonCameraScreen.java
@@ -9,13 +9,14 @@ import net.minecraft.resources.ResourceLocation;
 
 public class ThirdPersonCameraScreen extends SettingsScreenBase {
 
-    private static final ResourceLocation TEXTURE = new ResourceLocation(CameraUtils.MODID, "textures/gui/generic_6.png");
+    private static final ResourceLocation TEXTURE = new ResourceLocation(CameraUtils.MODID, "textures/gui/generic_7.png");
 
     private ConfigBuilder.ConfigEntry<Double> offsetX;
     private ConfigBuilder.ConfigEntry<Double> offsetY;
     private ConfigBuilder.ConfigEntry<Double> offsetZ;
     private ConfigBuilder.ConfigEntry<Double> rotationX;
     private ConfigBuilder.ConfigEntry<Boolean> inverted;
+    private ConfigBuilder.ConfigEntry<Boolean> hideGui;
     private int slot;
 
     public ThirdPersonCameraScreen(int slot,
@@ -23,15 +24,17 @@ public class ThirdPersonCameraScreen extends SettingsScreenBase {
                                    ConfigBuilder.ConfigEntry<Double> offsetY,
                                    ConfigBuilder.ConfigEntry<Double> offsetZ,
                                    ConfigBuilder.ConfigEntry<Double> rotationX,
-                                   ConfigBuilder.ConfigEntry<Boolean> inverted
+                                   ConfigBuilder.ConfigEntry<Boolean> inverted,
+                                   ConfigBuilder.ConfigEntry<Boolean> hideGui
     ) {
-        super(new TranslatableComponent("gui.camerautils.third_person_camera.title", slot + 1), TEXTURE, 248, 179);
+        super(new TranslatableComponent("gui.camerautils.third_person_camera.title", slot + 1), TEXTURE, 248, 204);
         this.slot = slot;
         this.offsetX = offsetX;
         this.offsetY = offsetY;
         this.offsetZ = offsetZ;
         this.rotationX = rotationX;
         this.inverted = inverted;
+        this.hideGui = hideGui;
     }
 
     @Override
@@ -70,7 +73,11 @@ public class ThirdPersonCameraScreen extends SettingsScreenBase {
                 inverted,
                 value -> new TranslatableComponent("message.camerautils.inverted", value)
         ));
-        addRenderableWidget(new Button(guiLeft + 10, guiTop + 7 + font.lineHeight + 10 + 25 * 5, xSize - 20, 20, new TranslatableComponent("message.camerautils.reset"), button -> {
+        addRenderableWidget(new ConfigValueButton(guiLeft + 10, guiTop + 7 + font.lineHeight + 10 + 25 * 5, xSize - 20, 20,
+                hideGui,
+                value -> new TranslatableComponent("message.camerautils.hide_gui", value)
+        ));
+        addRenderableWidget(new Button(guiLeft + 10, guiTop + 7 + font.lineHeight + 10 + 25 * 6, xSize - 20, 20, new TranslatableComponent("message.camerautils.reset"), button -> {
             offsetX.reset();
             offsetX.save();
             offsetY.reset();
@@ -81,7 +88,9 @@ public class ThirdPersonCameraScreen extends SettingsScreenBase {
             rotationX.save();
             inverted.reset();
             inverted.save();
-            minecraft.setScreen(new ThirdPersonCameraScreen(slot, offsetX, offsetY, offsetZ, rotationX, inverted));
+            hideGui.reset();
+            hideGui.save();
+            minecraft.setScreen(new ThirdPersonCameraScreen(slot, offsetX, offsetY, offsetZ, rotationX, inverted, hideGui));
         }));
     }